# all: optimize LineTo, CurveTo and methods formatting numbers