--- a/fpdf_test.go
+++ b/fpdf_test.go
@@ -2922,3 +2922,21 @@ func ExampleFpdf_SetModificationDate() {
 	// Output:
 	// Successfully generated pdf/Fpdf_SetModificationDate.pdf
 }
+
+func BenchmarkLineTo(b *testing.B) {
+	pdf := gofpdf.New("P", "mm", "A4", "")
+	pdf.AddPage()
+
+	for i := 0; i < b.N; i++ {
+		pdf.LineTo(170, 20)
+	}
+}
+
+func BenchmarkCurveTo(b *testing.B) {
+	pdf := gofpdf.New("P", "mm", "A4", "")
+	pdf.AddPage()
+
+	for i := 0; i < b.N; i++ {
+		pdf.CurveTo(190, 100, 105, 100)
+	}
+}